refactor: add store option
